Fully implement exclude feature
Polyglot now understands the commands "exclude|include MOVE|all".
The list of excluded moves is cleared on "undo" or any move, and can be
interrogated through the "bk" command (when book is off), printing an
asterisk in front of each excluded move. An "exclude all" is deferred
until there are again enabled moves. New searches are only started
when the list of excluded moves actually changes.