Repair Affinity option
Affinity was set in the engine startup routine, but the latter was always
done before the ini file had been completely read. The Affinity value
is now extracted explicitly from the ini file before engine startup.
The test for validity of the process handle was also reversed.
Affinity is marked as an XBoard option, but not in the standard set,
as it still would not work when set interactively. It does not work
in -noini mode, but requires an old-fashoned ini file.