Fix and improve move choice with learning
A new formula is used to recalculate weights when the BookLearn option is on.
Unlike the old one, this allows the effective weight to increase by an
arbitrarily large factor between a move that always wins and one that always
draws (or between one that always draws and one that always loses).